+++ cli_tool_audit/audit_manager.py
@@ -364,8 +364,9 @@ class AuditManager:
         # pylint: disable=broad-exception-caught
         try:
             command = [tool_name, version_switch]
+            timeout = int(os.environ.get("CLI_TOOL_AUDIT_TIMEOUT", 15))
             result = subprocess.run(
-                command, capture_output=True, text=True, timeout=15, shell=False, check=True
+                command, capture_output=True, text=True, timeout=timeout, shell=False, check=True
             )  # nosec
             # Sometimes version is on line 2 or later.
             version = result.stdout.strip()