+++ CHANGELOG.md
@@ -5,7 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.2.0] - 2024-01-19
+## [3.0.0] - 2024-01-19
+
+### Added
+- filter an audit by tags
+- install_docs, install_command to show what to do when there are problems
+- `single` to validate one tool without a config file
+- `.html` output
+
+## Fixed
+- only truncate long version on console output
+- removed three dependencies (toml, hypothesis, importlib-metadata)
+
+
+## [2.0.0] - 2024-01-19
 
 ### Fixed
 - Cache now adds gitingore and clears files older than 30 days on startup